[rhi] Reorganization
