+++ include/nova_renderer/frontend/rendergraph.hpp
@@ -1,0 +1,100 @@
+#pragma once
+
+#include "nova_renderer/frontend/procedural_mesh.hpp"
+#include "nova_renderer/renderables.hpp"
+#include "nova_renderer/rhi/rhi_types.hpp"
+#include "nova_renderer/util/container_accessor.hpp"
+
+namespace nova::renderer {
+    template <typename RenderCommandType>
+    struct MeshBatch {
+        rhi::Buffer* vertex_buffer = nullptr;
+        rhi::Buffer* index_buffer = nullptr;
+
+        /*!
+         * \brief A buffer to hold all the per-draw data
+         *
+         * For example, a non-animated mesh just needs a mat4 for its model matrix
+         *
+         * This buffer gets re-written to every frame, since the number of renderables in this mesh batch might have changed. If there's
+         * more renderables than the buffer can hold, it gets reallocated from the RHI
+         */
+        rhi::Buffer* per_renderable_data = nullptr;
+
+        std::vector<RenderCommandType> commands;
+    };
+
+    template <typename RenderCommandType>
+    struct ProceduralMeshBatch {
+        MapAccessor<MeshId, ProceduralMesh> mesh;
+
+        /*!
+         * \brief A buffer to hold all the per-draw data
+         *
+         * For example, a non-animated mesh just needs a mat4 for its model matrix
+         *
+         * This buffer gets re-written to every frame, since the number of renderables in this mesh batch might have changed. If there's
+         * more renderables than the buffer can hold, it gets reallocated from the RHI
+         */
+        rhi::Buffer* per_renderable_data = nullptr;
+
+        std::vector<RenderCommandType> commands;
+
+        ProceduralMeshBatch(std::unordered_map<MeshId, ProceduralMesh>* meshes, const MeshId key) : mesh(meshes, key) {}
+    };
+
+    struct MaterialPass {
+        // Descriptors for the material pass
+
+        std::vector<MeshBatch<StaticMeshRenderCommand>> static_mesh_draws;
+        std::vector<ProceduralMeshBatch<StaticMeshRenderCommand>> static_procedural_mesh_draws;
+
+        std::vector<rhi::DescriptorSet*> descriptor_sets;
+        const rhi::PipelineInterface* pipeline_interface = nullptr;
+    };
+
+    struct Pipeline {
+        rhi::Pipeline* pipeline = nullptr;
+
+        std::vector<MaterialPass> passes;
+    };
+
+    struct Renderpass {
+        uint32_t id = 0;
+
+        rhi::Renderpass* renderpass = nullptr;
+        rhi::Framebuffer* framebuffer = nullptr;
+
+        std::vector<Pipeline> pipelines;
+
+        bool writes_to_backbuffer = false;
+
+        std::vector<rhi::ResourceBarrier> read_texture_barriers;
+        std::vector<rhi::ResourceBarrier> write_texture_barriers;
+
+        /*!
+         * \brief Optional record function for this renderpass
+         *
+         * This member allows a host application to register a custom command list recording function for this renderpass. It paves the way
+         * for shaderpacks with render scripts
+         *
+         * The first parameter to this function is this exact renderpass, the second parameter is the command list to record into
+         */
+        std::optional<std::function<void(const Renderpass&, rhi::CommandList*)>> record;
+
+        /*!
+         * \brief Records this renderpass into the provided command list
+         *
+         * If this renderpass has a custom `record` function, this method will call that function. If not, it will simply render all the
+         * drawcalls in this renderpass
+         */
+        void record_into_command_list(rhi::CommandList* cmds);
+    };
+
+    struct Mesh {
+        rhi::Buffer* vertex_buffer = nullptr;
+        rhi::Buffer* index_buffer = nullptr;
+
+        uint32_t num_indices = 0;
+    };
+} // namespace nova::renderer
