+++ include/nova_renderer/nova_renderer.hpp
@@ -5,11 +5,13 @@
 #include <mutex>
 #include <string>
 
-#include "nova_renderer/device_memory_resource.hpp"
+#include "nova_renderer/frontend/rendergraph.hpp"
 #include "nova_renderer/nova_settings.hpp"
 #include "nova_renderer/polyalloc.hpp"
-#include "nova_renderer/render_engine.hpp"
 #include "nova_renderer/renderdoc_app.h"
+#include "nova_renderer/rhi/device_memory_resource.hpp"
+#include "nova_renderer/rhi/forward_decls.hpp"
+#include "nova_renderer/rhi/render_engine.hpp"
 
 #include "constants.hpp"
 #include "frontend/procedural_mesh.hpp"
@@ -27,79 +29,6 @@ namespace nova::renderer {
     }
 
 #pragma region Runtime optimized data
-    template <typename RenderCommandType>
-    struct MeshBatch {
-        rhi::Buffer* vertex_buffer = nullptr;
-        rhi::Buffer* index_buffer = nullptr;
-
-        /*!
-         * \brief A buffer to hold all the per-draw data
-         *
-         * For example, a non-animated mesh just needs a mat4 for its model matrix
-         *
-         * This buffer gets re-written to every frame, since the number of renderables in this mesh batch might have changed. If there's
-         * more renderables than the buffer can hold, it gets reallocated from the RHI
-         */
-        rhi::Buffer* per_renderable_data = nullptr;
-
-        std::vector<RenderCommandType> commands;
-    };
-
-    template <typename RenderCommandType>
-    struct ProceduralMeshBatch {
-        MapAccessor<MeshId, ProceduralMesh> mesh;
-
-        /*!
-         * \brief A buffer to hold all the per-draw data
-         *
-         * For example, a non-animated mesh just needs a mat4 for its model matrix
-         *
-         * This buffer gets re-written to every frame, since the number of renderables in this mesh batch might have changed. If there's
-         * more renderables than the buffer can hold, it gets reallocated from the RHI
-         */
-        rhi::Buffer* per_renderable_data = nullptr;
-
-        std::vector<RenderCommandType> commands;
-
-        ProceduralMeshBatch(std::unordered_map<MeshId, ProceduralMesh>* meshes, const MeshId key) : mesh(meshes, key) {}
-    };
-
-    struct MaterialPass {
-        // Descriptors for the material pass
-
-        std::vector<MeshBatch<StaticMeshRenderCommand>> static_mesh_draws;
-        std::vector<ProceduralMeshBatch<StaticMeshRenderCommand>> static_procedural_mesh_draws;
-
-        std::vector<rhi::DescriptorSet*> descriptor_sets;
-        const rhi::PipelineInterface* pipeline_interface = nullptr;
-    };
-
-    struct Pipeline {
-        rhi::Pipeline* pipeline = nullptr;
-
-        std::vector<MaterialPass> passes;
-    };
-
-    struct Renderpass {
-        uint32_t id = 0;
-
-        rhi::Renderpass* renderpass = nullptr;
-        rhi::Framebuffer* framebuffer = nullptr;
-
-        std::vector<Pipeline> pipelines;
-
-        bool writes_to_backbuffer = false;
-
-        std::vector<rhi::ResourceBarrier> read_texture_barriers;
-        std::vector<rhi::ResourceBarrier> write_texture_barriers;
-    };
-
-    struct Mesh {
-        rhi::Buffer* vertex_buffer = nullptr;
-        rhi::Buffer* index_buffer = nullptr;
-
-        uint32_t num_indices = 0;
-    };
 #pragma endregion
 
 #pragma region metadata
