+++ include/nova_renderer/rhi/device_memory_resource.hpp
@@ -1,6 +1,5 @@
 #pragma once
 
-#include "nova_renderer/polyalloc.hpp"
 #include "nova_renderer/allocation_structs.hpp"
 #include "nova_renderer/allocation_strategy.hpp"
 #include "nova_renderer/bytes.hpp"
