--- src/render_engine/command_list.cpp
@@ -1,1 +1,0 @@
-#include "nova_renderer/command_list.hpp"
