+++ src/render_engine/dx12/dx12_command_list.hpp
@@ -1,16 +1,10 @@
-/*!
- * \author ddubois
- * \date 30-Mar-19.
- */
-
-#ifndef NOVA_RENDERER_D3D12_COMMAND_LIST_HPP
-#define NOVA_RENDERER_D3D12_COMMAND_LIST_HPP
+#pragma once
 
 #pragma warning(push, 0)
 #include <d3d12.h>
 #pragma warning(pop)
 
-#include <nova_renderer/command_list.hpp>
+#include "nova_renderer/rhi/command_list.hpp"
 
 #pragma warning(push, 0)
 #include <wrl/client.h>
@@ -56,5 +50,3 @@ namespace nova::renderer::rhi {
         Microsoft::WRL::ComPtr<ID3D12Device> device;
     };
 } // namespace nova::renderer::rhi
-
-#endif // NOVA_RENDERER_D3D12_COMMAND_LIST_HPP
