+++ src/render_engine/render_engine.cpp
@@ -1,5 +1,5 @@
 #include <utility>
-#include "nova_renderer/render_engine.hpp"
+#include "nova_renderer/rhi/render_engine.hpp"
 
 namespace nova::renderer::rhi {
     void RenderEngine::set_shaderpack_data_allocator(const bvestl::polyalloc::allocator_handle& allocator_handle) {
