+++ src/render_objects/procedural_mesh.cpp
@@ -1,6 +1,7 @@
 #include "nova_renderer/frontend/procedural_mesh.hpp"
 
-#include "nova_renderer/render_engine.hpp"
+#include "nova_renderer/rhi/device_memory_resource.hpp"
+#include "nova_renderer/rhi/render_engine.hpp"
 
 #include "../memory/block_allocation_strategy.hpp"
 #include "../memory/mallocator.hpp"
